requirements.txt
Jinja2
ruamel.yaml
colorama
ansible==2.1.0.0
